# core/docs/swagger.yaml
basePath: /
definitions:
...
  datatypes.JSONType-db_GeoLocation:
    type: object
  db.Agent:
    properties:
      arch:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      kernel_version:
        type: string
      last_seen:
        type: string
      location:
        $ref: '#/definitions/datatypes.JSONType-db_GeoLocation'
      machine_id:
        type: string
      maintenance_mode:
        type: boolean
      name:
        type: string
      os:
        type: string
      platform:
        type: string
      reporting_interval_seconds:
        description: System metrics reporting interval
        type: integer
      token:
        type: string
    type: object
...
  service.RegisterRequest:
    properties:
      arch:
        type: string
      machine_id:
        type: string
      name:
        type: string
      os:
        type: string
    required:
    - arch
    - machine_id
    - name
    - os
    type: object
  service.RegisterResponse:
    properties:
      agent_id:
        type: string
      token:
        type: string
    type: object
  service.SetMaintenanceModeRequest:
    properties:
      maintenance_mode:
        type: boolean
    required:
    - maintenance_mode
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8999
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Orion Core API for agent management and monitoring
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Orion Core API
  version: "1.0"
paths:
...
  /v1/agents:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all registered agents
      parameters:
      - default: 50
        description: Maximum number of agents to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of agents to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  properties:
                    agents:
                      items:
                        $ref: '#/definitions/db.Agent'
                      type: array
                    count:
                      format: int64
                      type: integer
                    limit:
                      type: integer
                    offset:
                      type: integer
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: List agents
      tags:
      - agents
...
  /v1/agents/{agent_id}/maintenance:
    put:
      consumes:
      - application/json
      description: Enable or disable maintenance mode for a specific agent
      parameters:
      - description: Agent ID
        in: path
        name: agent_id
        required: true
        type: string
      - description: Maintenance mode request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.SetMaintenanceModeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Set agent maintenance mode
      tags:
      - agents
...
  /v1/agents/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific agent including latest
        report
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  properties:
                    agent:
                      $ref: '#/definitions/db.Agent'
                    latest_report:
                      type: object
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get agent details
      tags:
      - agents
...
  /v1/incidents/candidates:
    get:
      consumes:
      - application/json
      description: Get a list of monitors that are candidates for incidents based
        on health status and recent reports
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  properties:
                    candidates:
                      items:
                        type: object
                      type: array
                    count:
                      type: integer
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get incident candidates
      tags:
      - health
...
  /v1/register:
    post:
      consumes:
      - application/json
      description: Register a new agent or reconnect an existing agent by machine
        ID
      parameters:
      - description: Agent registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Register an agent
      tags:
      - agents
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"